% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{CurlR-package}
\alias{CurlR}
\alias{CurlR-package}
\title{CurlR}
\description{
Data package for Mixed Doubles Curling.
}
\author{
\strong{Maintainer}: Robyn Ritchie \email{your@email.com}

}
\keyword{internal}
