% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml2_action.R
\name{ele_action}
\alias{ele_action}
\alias{ele_click}
\alias{ele_fillText}
\alias{focus_first_window}
\alias{focus_last_window}
\title{Page element actions}
\usage{
ele_click(x)

ele_fillText(x, value, key = NULL, return = FALSE)

focus_first_window(p)

focus_last_window(p)
}
\arguments{
\item{key}{keyboard element, e.g. \code{enter}, \code{control}.}
}
\description{
Page element actions
}
\seealso{
https://cran.r-project.org/web/packages/RSelenium/vignettes/basics.html
}
