% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml2_find.R
\name{ele_find_first}
\alias{ele_find_first}
\alias{ele_find_first.remoteDriver}
\alias{ele_find_all}
\alias{ele_find_all.remoteDriver}
\alias{getElementById}
\alias{getElementByName}
\alias{getElementByClass}
\alias{ele_child}
\alias{ele_children}
\title{Search for an element on the page}
\usage{
ele_find_first(x, pattern, type = "xpath")

\method{ele_find_first}{remoteDriver}(x, pattern, type = "xpath")

ele_find_all(x, pattern, type = "xpath")

\method{ele_find_all}{remoteDriver}(x, value, type = "xpath")

getElementById(x, value)

getElementByName(x, value)

getElementByClass(x, value)

ele_child(x, value = "*", type = "xpath")

ele_children(x, value = "*", type = "xpath")
}
\arguments{
\item{x}{A remoteDriver object}

\item{type}{Locator scheme to use to search the element, available schemes:
\code{c("xpath", "class name", "css selector", "id", "name", "link text", "partial link text", "tag name")}. Defaults to 'xpath'. Partial string matching is
accepted.}

\item{value}{The search target.}
}
\description{
Search for an element on the page
}
\details{
Details of possible locator schemes
  \describe{
    \item{"class name" :}{Returns an element whose class name contains the search value; compound class names are not permitted.}
                                            \item{"css selector" :}{Returns an element matching a CSS selector.}
                                            \item{"id" :}{Returns an element whose ID attribute matches the search value.}
                                            \item{"name" :}{Returns an element whose NAME attribute matches the search value.}
                                            \item{"link text" :}{Returns an anchor element whose visible text matches the search value.}
                                            \item{"partial link text" :}{Returns an anchor element whose visible text partially matches the search value.}
                                            \item{"tag name" :}{Returns an element whose tag name matches the search value.}
                                            \item{"xpath" :}{Returns an element matching an XPath expression.}
                                          }
}
\seealso{
\code{\link[seleniumPipes:findElement]{seleniumPipes::findElement()}}
}
