% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendEmail.R
\name{EmailConfig_qq}
\alias{EmailConfig_qq}
\title{set the email information of the email sender}
\usage{
EmailConfig_qq(user, PassCode)
}
\arguments{
\item{user}{the email address of the sender}

\item{PassCode}{the pass code of the sender, the length is 16 for QQ \code{PassCode}}
}
\description{
Configuration will be write to \code{curlR/python/config.json}, where user, PassCode
and server will be stored.
}
\examples{
\dontrun{
EmailConfig_qq("991810576@qq.com", "XXXXXXXX")
}
}
