% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cookie.R
\name{GET2}
\alias{GET2}
\alias{POST2}
\title{request url with cookiefile}
\usage{
GET2(..., cookiefile = "cookies.txt")

POST2(..., cookiefile = "cookies.txt")
}
\arguments{
\item{...}{other parameters will be passed to \code{\link[=GET]{GET()}} or \code{\link[=POST]{POST()}}}

\item{cookiefile}{path of cookie file}
}
\description{
request url with cookiefile
}
