% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_selenium.R
\name{init_selenium}
\alias{init_selenium}
\alias{kill_selenium}
\title{init_selenium}
\usage{
init_selenium(port = 4444, browserName = "firefox")

kill_selenium(port = 4444)
}
\arguments{
\item{browserName}{firefox or chrome}
}
\description{
init_selenium
}
\examples{
# init_selenium(6666)
# kill_selenium(6666)
}
