% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mxd_games}
\alias{mxd_games}
\title{Mixed Doubles Curling — Game-Level Data}
\format{
A data frame with 1112 rows and 168 columns.
}
\source{
Created by Robyn Ritchie via text scraping and image processing of
  results books from \url{https://curlit.com/results}.
}
\usage{
mxd_games
}
\description{
A data frame containing game-level information for Mixed Doubles curling
events from OWG and WMDCC competitions.
}
\examples{
data("mxd_games")
str(mxd_games)
}
\keyword{datasets}
