% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mxd_shots}
\alias{mxd_shots}
\title{Mixed Doubles Curling — Shot-Level Data}
\format{
A data frame with 66632 rows and 104 columns.
}
\source{
Created by Robyn Ritchie via text scraping and image processing of
  results books from \url{https://curlit.com/results}.
}
\usage{
mxd_shots
}
\description{
A data frame containing shot-by-shot details from Mixed Doubles curling
competitions (OWG and WMDCC).
}
\examples{
data("mxd_shots")
str(mxd_shots)
}
\keyword{datasets}
