% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_email.R
\name{send_email}
\alias{send_email}
\title{send email by email and smtplib}
\usage{
send_email(
  receiver = "kongdd@mail2.sysu.edu.cn",
  title = "ChinaWater",
  content = ""
)
}
\description{
send email by email and smtplib
}
