% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendEmail.R
\name{sendEmail}
\alias{sendEmail}
\title{send email by email and smtplib}
\usage{
sendEmail(receiver = "kongdd@mail2.sysu.edu.cn", content = "", title = "curlR")
}
\arguments{
\item{receiver}{the email address of the receiver}

\item{title, content}{the mail title and content}
}
\description{
Please fill the configuration information first by the \code{\link[=EmailConfig_qq]{EmailConfig_qq()}}.
Config once, and user configuration will be stored.
}
\note{
currently only QQ email supported. But it is easy to extend more.
}
\examples{
\dontrun{
EmailConfig_qq("991810576@qq.com", "XXXXXXXX")
sendEmail("kongdd@mail2.sysu.edu.cn", content = "hello python email")
}
}
