% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sqlite.R
\name{write_db}
\alias{write_db}
\alias{write_db_char}
\alias{write_db_batch}
\alias{read_db}
\title{write_db}
\usage{
write_db(con, d, time, overwrite = FALSE, append = TRUE, mink = 10)

write_db_char(con, file, overwrite = FALSE, append = TRUE, mink = 10)

write_db_batch(dbname = "chinawater.db", files, overwrite = TRUE)

read_db(dbname, tables = NULL)
}
\arguments{
\item{time}{POSIXct or time character, e.g. "1998-01-01 01:00:00"}

\item{mink}{minimum records}
}
\description{
write_db
}
\examples{
# write_db_batch("chinawater.db", files)
}
