% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_quote.R
\name{write_quote}
\alias{write_quote}
\title{write_quote}
\usage{
write_quote(
  rscript,
  outfile,
 
    header = c("#! /usr/bin/Rscript \\nlibrary(ChinaWater); library(lubridate); print(getwd());")
)
}
\value{
file path of \code{rscript}
}
\description{
write_quote
}
