% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sqlite.R
\name{write_sqlite}
\alias{write_sqlite}
\alias{read_sqlite}
\title{Functions for writing data frames or delimiter-separated files to database tables.}
\usage{
write_sqlite(file, name, value, overwrite = FALSE, append = FALSE)

read_sqlite(file, name)
}
\arguments{
\item{file}{file path}

\item{name}{a character string specifying a table name. SQLite table names
are not case sensitive, e.g., table names ABC and abc are considered equal.}

\item{value}{a data.frame (or coercible to data.frame) object or a file
name (character). In the first case, the data.frame is written to a temporary
file and then imported to SQLite; when value is a character, it is interpreted
as a file name and its contents imported to SQLite.}

\item{overwrite}{a logical specifying whether to overwrite an existing table
or not. Its default is FALSE.}

\item{append}{a logical specifying whether to append to an existing table in the DBMS.
Its default is FALSE.}
}
\description{
Functions for writing data frames or delimiter-separated files to database tables.
}
