% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml2.R
\name{xml_html}
\alias{xml_html}
\alias{xml_html.webElement}
\alias{xml_html.list}
\title{xml_html}
\usage{
xml_html(x, parse = TRUE)

\method{xml_html}{webElement}(x, parse = TRUE)

\method{xml_html}{list}(x, parse = TRUE)
}
\description{
xml_html
}
